Dynamic inefficiency in pixel digitizer for FPIX

+++ SimTracker/SiPixelDigitizer/plugins/SiPixelDigitizerAlgorithm.h
@@ -240,14 +240,14 @@ class SiPixelDigitizerAlgorithm  {
    */
    struct PixelEfficiencies {
      PixelEfficiencies(const edm::ParameterSet& conf, bool AddPixelInefficiency, int NumberOfBarrelLayers, int NumberOfEndcapDisks);
-     float thePixelEfficiency[20];     // Single pixel effciency
-     float thePixelColEfficiency[20];  // Column effciency
-     float thePixelChipEfficiency[20]; // ROC efficiency
+     double thePixelEfficiency[20];     // Single pixel effciency
+     double thePixelColEfficiency[20];  // Column effciency
+     double thePixelChipEfficiency[20]; // ROC efficiency
      std::vector<double> theLadderEfficiency_BPix[20]; // Ladder efficiency
      std::vector<double> theModuleEfficiency_BPix[20]; // Module efficiency
      std::vector<double> thePUEfficiency[20]; // Instlumi dependent efficiency
-     float theInnerEfficiency_FPix[20]; // Fpix inner module efficiency
-     float theOuterEfficiency_FPix[20]; // Fpix outer module efficiency
+     double theInnerEfficiency_FPix[20]; // Fpix inner module efficiency
+     double theOuterEfficiency_FPix[20]; // Fpix outer module efficiency
      unsigned int FPixIndex;         // The Efficiency index for FPix Disks
    };
